cleanup - add license
