+++ TODO.md
@@ -1,3 +1,3 @@
-- jshint
 - optimize css and js with webpack
+- structure view files into includes / pages
 # Add these as github issues ?!
