+++ src/server/views/footer.pug
@@ -23,6 +23,7 @@ footer.text-center
 	link(href='https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css', rel='stylesheet', integrity='sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN', crossorigin='anonymous')
 	script(src='https://cdnjs.cloudflare.com/ajax/libs/jquery.isotope/3.0.4/isotope.pkgd.min.js')
 	script(src='https://cdnjs.cloudflare.com/ajax/libs/lightslider/1.1.6/js/lightslider.min.js')
+	script(src='scripts/sha512.js')
 	script(src='scripts/helper.js')
 	script(src='scripts/js-cookie.js')
 	script(src='scripts/main.js')
@@ -31,5 +32,5 @@ footer.text-center
 		(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
 		m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
 		})(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-		ga('create', 'UA-62655804-1', 'auto');
+		ga('create', env.GOOGLE_ANALYTICS, 'auto');
 		ga('send', 'pageview');
