+++ src/server/views/header.pug
@@ -2,14 +2,10 @@ doctype html
 html
   head
     title NetsBlox
-    //- link(href='libs/bootstrap/bootstrap-3.1.1-dist/css/bootstrap.min.css', rel='stylesheet')
-    //- link(href='libs/bootstrap/bootstrap-3.1.1-dist/css/bootstrap-theme.min.css', rel='stylesheet')
     link(rel='stylesheet', href='https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css', integrity='sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u', crossorigin='anonymous')
     link(href='style/carousel_theme.css', rel='stylesheet')
-    //- link(href='libs/rrssb/rrssb.css', rel='stylesheet')
     link(href='style/webgme_www.css', rel='stylesheet')
     link(href='https://cdnjs.cloudflare.com/ajax/libs/lightslider/1.1.6/css/lightslider.min.css', rel='stylesheet')
-    //- link(rel='stylesheet', href='libs/fancybox/jquery.fancybox.css', type='text/css', media='screen')
     link(rel='stylesheet', type='text/css', href='style/main.css')
     meta(name='viewport', content='width=device-width, initial-scale=1')
     meta(name='description', content='NetsBlox is a visual programming language and cloud-based environment that enables novice programmers to create networked programs such as multi-player games.')
