+++ src/server/views/pubProjects.pug
@@ -1,17 +1,3 @@
-//- .hidden-xs
-//-   h3.text-center Published Projects
-//-   ul#projects-slider.cS-hidden
-//-     each proj in projects
-//-       .prj-element.element-item
-//-         a(href=env.EDITOR_ADDRESS + '/#present:Username='+proj.owner+'&ProjectName='+proj.projectName, target='_blank')
-//-           .thumbnail
-//-             img(src= proj.thumbnail, alt='NetsBlox Project: ' + proj.projectName)
-//-             .caption.text-center
-//-               h4= proj.projectName
-//-               small= proj.services[0]
-
-
-
 .hidden-xs
   h3.text-center Published Projects
   ul#projects-slider.cS-hidden
