+++ src/server/views/tutorials.pug
@@ -24,29 +24,3 @@ block content
 					.panel-body
 						.embed-responsive.embed-responsive-16by9
 							iframe(class="embed-responsive-item", width='100%', height='400px', frameborder='0', allowfullscreen='', src='https://www.youtube.com/embed/-qS7hGowQKQ')
-
-
-
-
-
-
-
-
-
-
-							//- .container.text-center
-							//- 	.row
-							//- 		.col-md-6
-							//- 			.panel.panel-default
-							//- 				.panel-body
-							//- 					.embed-responsive.embed-responsive-16-9
-							//- 						iframe(class="embed-responsive-item", width='100%', height='400px', frameborder='0', allowfullscreen='', src='https://www.youtube.com/embed/YXKUtTbbVWI')
-							//- 		.col-md-6
-							//- 			.panel.panel-default
-							//- 				.panel-body
-							//- 					iframe(width='100%', height='400px', frameborder='0', allowfullscreen='', src='https://www.youtube.com/embed/NRIUUiJMCmI')
-							//- 	.row
-							//- 		.col-md-6
-							//- 			.panel.panel-default
-							//- 				.panel-body
-							//- 					iframe(width='100%', height='400px', frameborder='0', allowfullscreen='', src='https://www.youtube.com/embed/-qS7hGowQKQ')
